requirements.txt
fastapi==0.115.6
ffmpeg_python==0.2.0
numpy== 1.26.2  
soundfile==0.13.0
starlette==0.41.3   
transformers==4.47.0
fastapi==0.115.6
scipy==1.15.1
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
noisereduce==3.0.3
silero-vad==5.1.2
pyjwt==2.10.1
noisereduce==3.0.3
silero-vad==5.1.2
pyjwt==2.10.1
python-multipart==0.0.20

